+++ astroquery/xmatch/core.py
@@ -146,7 +146,12 @@ class XMatchClass(BaseQuery):
 
         if not self.is_table_available(cat):
             if ((colRA is None) or (colDec is None)):
-                raise ValueError('Specify the name of the RA/Dec columns in the input table.')
+                raise ValueError(
+                    f"'{cat}' is not available on the XMatch server.If you are "
+                    "using a VizieR table name, note that only tables with "
+                    "coordinates are available on the XMatch server. If you are "
+                    f"using a local table, the arguments 'colRA{cat_index}' and "
+                    f"'colDec{cat_index}' must be provided.")
             # if `cat1` is not a VizieR table,
             # it is assumed it's either a URL or an uploaded table
             payload['colRA{0}'.format(cat_index)] = colRA